% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_news_files.R
\name{annotate_news_files}
\alias{annotate_news_files}
\title{Read NEWS.md files for tidyverse pkgs}
\usage{
annotate_news_files(files, dir = "news-files-annotated", quiet = TRUE)
}
\arguments{
\item{files}{A named list - names indicate packages and values indicate
filepaths of NEWS.md files}

\item{dir}{The directory to write the files to}
}
\value{
A named list of filepaths to the created documents
}
\description{
This function uses pandoc to read and parse NEWS.md files for tidyverse
packages. A custom lua filter is used to add annotations indicating lines
where individual bullets and bullet lists begin/end.
}
