% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_segmentable}
\alias{is_segmentable}
\title{is_segmentable("hello there!", width = 5, split = " ")
is_segmentable("hello there!", width = 5, split = " ", first_only = FALSE)}
\usage{
is_segmentable(x, width = 220, split = "\\n", first_only = TRUE)
}
\description{
is_segmentable("hello there!", width = 5, split = " ")
is_segmentable("hello there!", width = 5, split = " ", first_only = FALSE)
}
